requirements.txt
Django
djangorestframework
pandas
scikit-learn
gunicorn